Move model from experimental to demos folder
