+++ .github/CODEOWNERS
@@ -199,7 +199,7 @@ models/perf/ @yieldthought @uaydonat
 models/perf/benchmarking_utils.py @skhorasganiTT @williamlyTT @uaydonat
 models/demos/utils/llm_demo_utils.py @skhorasganiTT @mtairum @uaydonat
 models/experimental/stable_diffusion_xl_base @mbezuljTT @pavlepopovic @ipotkonjak-tt @tenstorrent/metalium-developers-convolutions
-models/experimental/mobilenetv2 @tenstorrent/cse-developer-ttnn
+models/demos/mobilenetv2 @tenstorrent/cse-developer-ttnn
 models/experimental/yolov8s_world @tenstorrent/cse-developer-ttnn
 models/experimental/yolov10 @tenstorrent/cse-developer-ttnn
 models/experimental/functional_yolov9c @tenstorrent/cse-developer-ttnn
