+++ models/demos/mobilenetv2/tests/mobilenetv2_e2e_performant.py
@@ -3,7 +3,7 @@
 # SPDX-License-Identifier: Apache-2.0
 
 import ttnn
-from models.experimental.mobilenetv2.tests.mobilenetv2_test_infra import create_test_infra
+from models.demos.mobilenetv2.tests.mobilenetv2_test_infra import create_test_infra
 
 try:
     pass
