+++ models/demos/mobilenetv2/tests/test_e2e_performant.py
@@ -9,7 +9,7 @@ import torch
 from loguru import logger
 
 from models.utility_functions import run_for_wormhole_b0
-from models.experimental.mobilenetv2.tests.mobilenetv2_e2e_performant import MobileNetV2Trace2CQ
+from models.demos.mobilenetv2.tests.mobilenetv2_e2e_performant import MobileNetV2Trace2CQ
 
 
 @run_for_wormhole_b0()
