+++ models/demos/mobilenetv2/tt/common.py
