+++ tests/ttnn/integration_tests/mobilenetv2/test_mobilenetv2.py
@@ -7,12 +7,12 @@ import os
 import torch
 import pytest
 import ttnn
-from models.experimental.mobilenetv2.reference.mobilenetv2 import Mobilenetv2
-from models.experimental.mobilenetv2.tt.model_preprocessing import (
+from models.demos.mobilenetv2.reference.mobilenetv2 import Mobilenetv2
+from models.demos.mobilenetv2.tt.model_preprocessing import (
     create_mobilenetv2_input_tensors,
     create_mobilenetv2_model_parameters,
 )
-from models.experimental.mobilenetv2.tt import ttnn_mobilenetv2
+from models.demos.mobilenetv2.tt import ttnn_mobilenetv2
 from tests.ttnn.utils_for_testing import assert_with_pcc
 
 
@@ -34,9 +34,9 @@ from tests.ttnn.utils_for_testing import assert_with_pcc
 )
 def test_mobilenetv2(device, use_pretrained_weight, batch_size, reset_seeds):
     # Check if weights file exists, if not, download them
-    weights_path = "models/experimental/mobilenetv2/mobilenet_v2-b0353104.pth"
+    weights_path = "models/demos/mobilenetv2/mobilenet_v2-b0353104.pth"
     if not os.path.exists(weights_path):
-        os.system("bash models/experimental/mobilenetv2/weights_download.sh")
+        os.system("bash models/demos/mobilenetv2/weights_download.sh")
     if use_pretrained_weight:
         state_dict = torch.load(weights_path)
         ds_state_dict = {k: v for k, v in state_dict.items()}
